Small ui fixes
