+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/di/modules/NetModule.java
@@ -1,0 +1,34 @@
+package sasd97.java_blog.xyz.yandexweather.di.modules;
+
+import javax.inject.Singleton;
+
+import dagger.Module;
+import dagger.Provides;
+import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
+import retrofit2.converter.gson.GsonConverterFactory;
+import sasd97.java_blog.xyz.yandexweather.data.net.WeatherApi;
+
+/**
+ * Created by alexander on 13/07/2017.
+ */
+
+@Module
+public class NetModule {
+
+    @Provides
+    @Singleton
+    public Retrofit provideRetrofit() {
+        return new Retrofit.Builder()
+                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+                .addConverterFactory(GsonConverterFactory.create())
+                .baseUrl("http://api.openweathermap.org/data/2.5/")
+                .build();
+    }
+
+    @Provides
+    @Singleton
+    public WeatherApi provideApi(Retrofit retrofit) {
+        return retrofit.create(WeatherApi.class);
+    }
+}
