+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/presentation/main/MainView.java
@@ -8,6 +8,5 @@ import com.arellomobile.mvp.MvpView;
 
 public interface MainView extends MvpView {
     void closeDrawer();
-    void updateToolbar(String title);
     void selectNavigationItem(int id);
 }
